Interfaces classes and adapter classes implemented
