--- src/main/java/com/example/sudokuminiproyecto2fpoe/model/Board.java
@@ -1,4 +1,0 @@
-package com.example.sudokuminiproyecto2fpoe.model;
-
-public class Board {
-}
