+++ src/main/java/com/example/sudokuminiproyecto2fpoe/view/iHomeStage.java
@@ -1,0 +1,7 @@
+package com.example.sudokuminiproyecto2fpoe.view;
+
+public interface iHomeStage {
+
+    void initStage();
+
+}
